requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Babel==2.11.0
bcrypt==4.1.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
customtkinter==5.2.0
darkdetect==0.8.0
distlib==0.3.7
django-crispy-forms==2.0
django-uni-form==0.9.0
et-xmlfile==1.1.0
filelock==3.12.4
Flask==3.0.0
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mssql==1.0.1
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.6
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.3
tk==0.1.0
tkcalendar==1.6.1
ttkbootstrap==1.10.1
ttkthemes==3.2.2
ttkwidgets==0.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uuid==1.30
virtualenv==20.24.5
Werkzeug==3.0.0
xlrd==2.0.1
